Fix: kernel channel destroy on consumer if sent
An enable-channel immediately followed by a destroy session triggered an
assert on the consumer followed by the shutdown of the consumer thread
for the kernel.

This fix make sure we don't ask the consumer to destroy a channel if not
previously sent to it and handle not found channel on destroy.

Fixes #643

Signed-off-by: David Goulet <dgoulet@efficios.com>

Conflicts:
	src/common/kernel-consumer/kernel-consumer.c
	- No health in stable-2.3 thus had to remove a health update
	code statement.