Rename and improve lttng conf file API
Remove the need of a path for each command since it can be done
automatically by looking up the HOME directory inside the conf API.

Rename namespace to "conf_*" since config_* is now used by libconfig.

This is an overall improvement of the code so we can expand it with
things that makes more sense.

Signed-off-by: David Goulet <dgoulet@efficios.com>