Fix: don't send consumer error to sessiond in teardown
At that point, the control socket has been closed thus the session
daemon is aware that the consumer died. The consumer exit reason is not
even handled by the session daemon.

In the case of the session daemon tearing down, the consumer should
avoid trying to send data on a potentially closed socket.

Signed-off-by: David Goulet <dgoulet@efficios.com>