Fix: auto load session in the auto/ directory
The session daemon will now automatically load sessions from the auto/
directory in the session config path. Both system config path and home
config path are used. The session is only loaded if the UID of the
session file matches the UID of the session daemon (or root).

For that, a autoload parameter has been added to the
config_load_session() function to indicate that we want to auto load
sessions.

Fixes #812

Signed-off-by: David Goulet <dgoulet@efficios.com>