Fix: validate file path creds in autoload mode
Validate the credentials only when in auto load mode for the auto/
directory. The lttng load command now checks the session file creds and
validate if it's readable before trying to do so.

Thus with this, a user can load a session from the system wide directory
as long as she/he has the rights to read it.

Signed-off-by: David Goulet <dgoulet@efficios.com>