Fix: disable event for JUL domain
This fix could not be backported from master since it has changed too
much but the ref. is 0dcfcf94b356c12e3936fa27f350896d96724e4e on the
master branch.

This fixes the disable event command that was not working with the JUL
domain and adds a test for it.

Signed-off-by: David Goulet <dgoulet@efficios.com>