Tests: Java agent: update after Java agent refactoring
This patch fixes the Java agent tests which were impossible to run since
the Jar files have been renamed. It also overhauls the java part of the
configure script to match what's in UST and replace some hardcoded path
with configurable values.

The tests are now enabled when corresponding classes are found in the
CLASSPATH during configure.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>