Send data pending status as part of payload instead of an invalid error
This ensures that the session daemon's logs don't contain false
positives every time a data pending command is replied-to.

The use of invalid error codes 0 and 1 causes the log to contain
misleading entries of the form:

Sending response (size: 16, retcode: Unknown error code (0)) [...]

This commit also proves that it is possible, contrary to what the
previous comment indicated, to send a boolean value over a socket
without using undocumented error codes.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>