Redirect spawned sessiond's stdout/err to /dev/null
Since the session daemon forked by `lttng create` shares its
standard output/error FDs, redirecting the standard output of
this command to another program "hangs" because the session daemon
never terminates.

Example that's not working (when sessiond is not running):

    lttng --mi xml create | xmllint --format -

This patch makes use of dup2() to redirect sessiond's standard output
and error to /dev/null.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>